feat: T25.5 Progress Dashboard CU-008 y US25 con pruebas

+++ backend/app/Domains/Workflow/Services/ProgressCalculationService.php
@@ -50,6 +50,6 @@ class ProgressCalculationService
             ]);
         }
 
-        return round($normalizedProgress, 2);
+        return (float) round($normalizedProgress, 2);
     }
 }
